+++ honeycomb-core/src/dart.rs
@@ -10,11 +10,9 @@
 
 // ------ IMPORTS
 
-// ------ CONTENT
-
-use std::sync::atomic::AtomicBool;
-
 use crate::{FaceIdentifier, VertexIdentifier};
+
+// ------ CONTENT
 
 /// Type definition for dart identifiers
 ///
@@ -54,50 +52,18 @@ pub struct CellIdentifiers {
 /// Structure used to store dart-related data. The association of data with
 /// a given dart is done implicitly through indexing.
 ///
-/// # Generics
-///
-/// - `const N_MARKS: usize` -- Number of marks used for search algorithms.
-///   This corresponds to the number of search that can be done concurrently.
-///
 /// # Example
 ///
 /// No example is provided as the structure should not be used directly.
 /// The documentation is generated mostly for developing purposes.
 ///
-pub struct DartData<const N_MARKS: usize> {
-    /// Array of boolean used for algorithmic search.
-    ///
-    /// Atomics allow for non-mutable interfaces, i.e. parallel friendly
-    /// methods. Storage is done line-wise as it would be very rare to
-    /// access multiple marks of the same dart successively.
-    pub marks: [Vec<AtomicBool>; N_MARKS],
+#[cfg_attr(feature = "benchmarking_utils", derive(Clone))]
+pub struct DartData {
     /// List of associated cell identifiers.
     pub associated_cells: Vec<CellIdentifiers>,
 }
 
-#[cfg(feature = "benchmarking_utils")]
-impl<const N_MARKS: usize> Clone for DartData<N_MARKS> {
-    fn clone(&self) -> Self {
-        Self {
-            marks: self
-                .marks
-                .iter()
-                .map(|elem| {
-                    elem.iter()
-                        .map(|atombool| {
-                            AtomicBool::new(atombool.load(std::sync::atomic::Ordering::Relaxed))
-                        })
-                        .collect::<Vec<AtomicBool>>()
-                })
-                .collect::<Vec<Vec<AtomicBool>>>()
-                .try_into()
-                .unwrap(),
-            associated_cells: self.associated_cells.clone(),
-        }
-    }
-}
-
-impl<const N_MARKS: usize> DartData<N_MARKS> {
+impl DartData {
     /// Create a [DartData] object.
     ///
     /// **This should not be used directly by the user.**
@@ -111,17 +77,7 @@ impl<const N_MARKS: usize> DartData<N_MARKS> {
     /// Returns a [DartData] structure with default values.
     ///
     pub fn new(n_darts: usize) -> Self {
-        let marks: [Vec<AtomicBool>; N_MARKS] = (0..N_MARKS)
-            .map(|_| {
-                (0..n_darts + 1)
-                    .map(|_| AtomicBool::new(false))
-                    .collect::<Vec<AtomicBool>>()
-            })
-            .collect::<Vec<Vec<AtomicBool>>>()
-            .try_into()
-            .unwrap();
         Self {
-            marks,
             associated_cells: vec![
                 CellIdentifiers {
                     vertex_id: 0,
@@ -133,62 +89,11 @@ impl<const N_MARKS: usize> DartData<N_MARKS> {
         }
     }
 
-    /// Free all darts of a given mark.
-    ///
-    /// **This should not be used directly by the user.**
-    ///
-    /// # Arguments
-    ///
-    /// - `mark_id: usize` -- Identifier of the mark that must be reset.
-    ///
-    pub fn free_mark(&self, mark_id: usize) {
-        self.marks[mark_id]
-            .iter()
-            .filter(|e| e.load(std::sync::atomic::Ordering::Relaxed)) // useful?
-            .for_each(|mark| mark.store(false, std::sync::atomic::Ordering::Relaxed));
-    }
-
-    /// Check and update if a given dart was marked.
-    ///
-    ///  # Arguments
-    ///
-    /// - `mark_id: usize` -- Identifier of the mark that must be checked.
-    /// - `dart_id: DartIdentifier` -- Identifier of the dart that must be checked.
-    ///
-    /// # Return / Panic
-    ///
-    /// Returns a boolean to indicate whether the dart was marked or not. In
-    /// both case, the dart is marked after the operation.
-    ///
-    /// The method will panic if the provided mark ID is invalid.
-    ///
-    pub fn was_marked(&self, mark_id: usize, dart_id: DartIdentifier) -> bool {
-        assert!(mark_id < N_MARKS);
-        match self.marks[mark_id][dart_id as usize].compare_exchange(
-            false,
-            true,
-            std::sync::atomic::Ordering::Release,
-            std::sync::atomic::Ordering::Relaxed,
-        ) {
-            Ok(_) => {
-                // comparison was successful => was false
-                false
-            }
-            Err(_) => {
-                // comparison failed => was true
-                true
-            }
-        }
-    }
-
     /// Add a new entry to the structure.
     ///
     /// **This should not be used directly by the user.**
     ///
     pub fn add_entry(&mut self) {
-        self.marks
-            .iter_mut()
-            .for_each(|mark| mark.push(AtomicBool::new(false)));
         self.associated_cells.push(CellIdentifiers::default());
     }
 
@@ -197,9 +102,6 @@ impl<const N_MARKS: usize> DartData<N_MARKS> {
     /// **This should not be used directly by the user.**
     ///
     pub fn add_entries(&mut self, n_darts: usize) {
-        self.marks
-            .iter_mut()
-            .for_each(|mark| mark.extend((0..n_darts).map(|_| AtomicBool::new(false))));
         self.associated_cells
             .extend((0..n_darts).map(|_| CellIdentifiers::default()));
     }
@@ -213,9 +115,6 @@ impl<const N_MARKS: usize> DartData<N_MARKS> {
     /// - `dart_id: DartIdentifier` -- Identifier of the dart which entry should be reset.
     ///
     pub fn reset_entry(&mut self, dart_id: DartIdentifier) {
-        self.marks.iter().for_each(|mark| {
-            mark[dart_id as usize].store(false, std::sync::atomic::Ordering::Relaxed)
-        });
         self.associated_cells[dart_id as usize] = CellIdentifiers::default();
     }
 }